A bunch of changes including Camdram single-sign-on integration, and a new User model. Also first two controllers (Home and Session, the latter of which deals with login via Camdram). Now depends on "haml" and "FakeWeb" gems.
git-svn-id: http://romeo.adctheatre.com/svn/archives/trunk@11 b587572d-e64b-43eb-89e5-c2811ed16773

+++ lib/tasks/setup.rake
@@ -1,7 +1,8 @@
 desc "Setup environment for ADC archives project"
 task :setup do
-  puts "Copying config/database.yml.example to config/database.yml."
-  puts "You can edit config/database.yml to customize your database settings."
-  FileUtils.cp("config/database.yml.example", "config/database.yml")
+  unless File.exist? "config/database.yml"
+    puts "Copying config/database.yml.example to config/database.yml."
+    FileUtils.cp("config/database.yml.example", "config/database.yml")
+  end
 end
